add delete operations in views and search student

+++ App/controllers/review.py
@@ -7,9 +7,11 @@ def create_review (studentID, staffID, experience, rating):
     db.session.commit()
 
 def delete_review(reviewID):
-    review= Review.query.get(reviewID)
-    db.session.delete(review)
-    db.session.commit()
+    review= Review.query.filter_by(reviewID=reviewID).first()
+    if review:
+        db.session.delete(review)
+        db.session.commit()
+    return review
 
 def get_review(reviewID):
     return Review.query.get(reviewID)
